% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{db.importing}
\alias{db.importing}
\title{Alias to \code{yload_dfx}}
\usage{
db.importing(
  li = NULL,
  frm = INPUTROOT,
  pattern = NULL,
  fextname = NULL,
  worker = NULL,
  retmode = "local",
  verbose = TRUE,
  row.names = NULL,
  sheet_name = 1,
  ...
)
}
\description{
Alias to \code{yload_dfx}
}
\seealso{
\code{\link{yload_dfx}}
}
