% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.ydumpto.function}
\alias{.ydumpto.function}
\title{Title}
\usage{
\method{.ydumpto}{`function`}(x, args, fpath, ffname, fextname, verbose = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
return plot_func and its args used for plotting named 'plot_func' & 'plot_args'
args is the list of args used by the plotting function
}
