% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{echo}
\alias{echo}
\title{display a dataframe/matrix table with more rows and columns displayed}
\usage{
echo(mat, nrowmax = 100, ncolmax = 100)
}
\arguments{
\item{mat}{the dataframe/matrix to be displayed}

\item{nrowmax}{the max number of rows to display}

\item{ncolmax}{the max number of columns to display}
}
\description{
used with Jupyter notebook display of data table, Need View function available at Global.Env
}
\examples{

\dontrun{
 data(mtcars)
 echo(mtcars)
}


}
