% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_na}
\alias{fill_na}
\title{fill NA values in a mat, prefer to be used with pipe %>% operators}
\usage{
fill_na(mat, value)
}
\arguments{
\item{mat}{a matrix or data.frame}

\item{value}{value to replace NA values in the mat}
}
\value{
mat with NA filled with value
}
\description{
can be easily placed in a pipe line %>%
}
