% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globals.R
\docType{data}
\name{INPUTROOT}
\alias{INPUTROOT}
\title{Input root directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
INPUTROOT
}
\description{
This variable defines the root directory for input files.
}
\keyword{datasets}
