% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.custom}
\alias{make.custom}
\title{Set the size for the following plots}
\usage{
make.custom(w = NULL, h = NULL, .set_global = TRUE)
}
\arguments{
\item{w}{set global WIDHT,  in inch, if length(w) == 2}

\item{h}{set global HEIGHT, in inch}

\item{.set_global}{TRUE, set global vars, else no global variables set}
}
\value{
Nothing is returned
}
\description{
basically, this function affects the following plot function which will use global var WIDTH and HEIGHT
  to determine the display repr.plot.width and repr.plot.height, set global variables depend on the third parameter
  `.set_global`
}
