% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globals.R
\docType{data}
\name{OUTPUTROOT}
\alias{OUTPUTROOT}
\title{Output root directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
OUTPUTROOT
}
\description{
This variable defines the root directory for output files.
}
\keyword{datasets}
