% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_anno_col}
\alias{show_anno_col}
\title{show_anno_col}
\usage{
show_anno_col(col_anno, ...)
}
\arguments{
\item{...}{allowed value are c(labels = FALSE,borders = FALSE,cex_label = ?)}
}
\description{
show colors for COL_ANNO
}
