% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globals.R
\docType{data}
\name{WIDTH}
\alias{WIDTH}
\title{use to control the default output fig width}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WIDTH
}
\description{
This variable defines the default width of the output figs, commonly used by
  function ysave, ydumpto. like cairo_pdf(..., height = HEIGHT, width = WIDTH)
}
\keyword{datasets}
