% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{ydeframe}
\alias{ydeframe}
\title{Convert column named data.frame to named list}
\usage{
ydeframe(df)
}
\arguments{
\item{df}{a data.frame object}
}
\value{
list
}
\description{
Convert a data.frame into list in length of ncol(df), with names(list) is the df's colnames
and each
}
