% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{ydo_count_deseq2}
\alias{ydo_count_deseq2}
\title{Title}
\usage{
ydo_count_deseq2(
  cnt,
  colData,
  col.id = "Tumor_Sample_Barcode",
  col.group = "Clin_classification",
  levels = NULL,
  auto.filter = TRUE,
  mode = "auto"
)
}
\arguments{
\item{mode}{}
}
\description{
Title
}
