% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{ydo_count_normlization_deseq2}
\alias{ydo_count_normlization_deseq2}
\title{A data.frame
The “stat” column in DESeq2 different expression analysis output refers to the Wald test statistic. The Wald test is commonly used for hypothesis testing when comparing two groups. And it is determined by the log2FoldChange and pvalue column values which makes it a good ranking index for further order based analysis like GSEA pre-ranked procedure.
The DESeq2 different expression analysis output contains several columns. The “baseMean” column contains the mean of normalized counts for each gene across all samples1. The “log2FoldChange” column contains the log2 fold change between two groups1. The “lfcSE” column contains the standard error of the log2 fold change1. The “stat” column contains the Wald test statistic2. The “pvalue” column contains the p-value of the Wald test2. The “padj” column contains the adjusted p-value for multiple testing correction3.}
\usage{
ydo_count_normlization_deseq2(cnt.sorted, colData.sorted)
}
\description{
A data.frame
The “stat” column in DESeq2 different expression analysis output refers to the Wald test statistic. The Wald test is commonly used for hypothesis testing when comparing two groups. And it is determined by the log2FoldChange and pvalue column values which makes it a good ranking index for further order based analysis like GSEA pre-ranked procedure.
The DESeq2 different expression analysis output contains several columns. The “baseMean” column contains the mean of normalized counts for each gene across all samples1. The “log2FoldChange” column contains the log2 fold change between two groups1. The “lfcSE” column contains the standard error of the log2 fold change1. The “stat” column contains the Wald test statistic2. The “pvalue” column contains the p-value of the Wald test2. The “padj” column contains the adjusted p-value for multiple testing correction3.
}
