% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maftools.R
\name{ydo_extract_cosmic_signatures_related_all}
\alias{ydo_extract_cosmic_signatures_related_all}
\title{Extract cosmic signatures from Maf object}
\usage{
ydo_extract_cosmic_signatures_related_all(
  laml,
  cosmic_ver = "v2",
  rank = 3,
  nrun = 3,
  single_core = TRUE,
  ...
)
}
\arguments{
\item{laml}{maftools::maf object}

\item{cosmic_ver}{c('v2','v3')}

\item{rank}{number of signatures extracted}

\item{nrun}{resample times}

\item{single_core}{[MutationalPatterns::extract_signatures()]}

\item{...}{pass to}
}
\description{
Extract cosmic signatures from Maf object
}
