% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ydo_GO}
\alias{ydo_GO}
\title{do GO analysis of a vector of gene symbols}
\usage{
ydo_GO(
  sig_genes,
  all_genes = NULL,
  ont = "ALL",
  p.t = 0.01,
  p.adjust.t = 0.05,
  ggsci_col_scheme = NULL,
  ...
)
}
\arguments{
\item{sig_genes}{sig_genes是all_genes的子集,sig_genes与all_genes的占比决定了GO富集的结果}

\item{all_genes}{all_genes必须包含全部基因(as ALL)而不是基因子集,GO分析才准确(https://mp.weixin.qq.com/s/zxiMXqRriGHcGAlCNk0e7A),
genes(格式为基因正式的名称symbols,不是ID)不为空,使用genes/ALL进行GO分析, 如果NULL,使用}

\item{ont}{ont is Ontology, one of c('BP', 'MF', 'CC', 'ALL')
ALL会进行全部3个富集分析:'BP', 'MF', 'CC'}

\item{ggsci_col_scheme}{默认使用pal_npg()(10)[c(3,6,9)]来标识BP,MF,CC的富集结果}

\item{...}{其他传入enrichGO函数的参数}
}
\value{
a list obj of class 'y.GO.res', used for input for yplot_GO_res
}
\description{
do GO analysis of a vector of gene symbols
}
