% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ydo_GO_bydiff}
\alias{ydo_GO_bydiff}
\title{Do GO analysis using different expression results}
\usage{
ydo_GO_bydiff(
  diff,
  ont = "ALL",
  by.expr.direction = FALSE,
  go.tb.filter = "top10",
  pvalueCutoff = 0.2,
  p.adjust.t = 0.05,
  t.p.val = NA,
  t.log.fc = 1,
  t.p.adj = 0.05,
  ...
)
}
\arguments{
\item{diff}{a data.frame with columns: symbol, log2FC, pvalue, padj, stat}

\item{ont}{Ontology, in c('BP', 'MF', 'CC', 'ALL')}

\item{by.expr.direction}{=TRUE, if TRUE genes in diff is group_by change direction before doing
GO analysis}

\item{p.adjust.t}{padj threshold, default 0.05}

\item{t.p.adj}{}
}
\value{
a list obj of class 'y.GO.res', used for input for yplot_GO_res
}
\description{
Do GO analysis using different expression results
}
\details{
如果genes is.null, use p.t or p.adjust.t && log2fc.abs.t (as FILTERS) to filter diff to
  get genes ' diff必须包含全部基因(as
  ALL)而不是基因子集,GO分析才准确(https://mp.weixin.qq.com/s/zxiMXqRriGHcGAlCNk0e7A) '
  若genes(格式为symbols)不为空,使用genes/ALL进行GO分析,若genes为空, 使用FILTERS从diff中筛选出genes"
}
