% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ydo_GSEA}
\alias{ydo_GSEA}
\title{GSEA analysis using different expression results}
\usage{
ydo_GSEA(
  diff = NULL,
  gene_list = NULL,
  set = "gsea",
  p.adjust.t = 0.05,
  p.t = NA,
  minGSSize = 10,
  maxGSSize = 500,
  .lineheight = 0.75,
  pvalueCutoff = 0.2,
  ...
)
}
\arguments{
\item{diff}{a data.frame which is a output df of DESeq2 different expr analysis output, with
columns: symbol, log2FC, pvalue, padj, stat}

\item{gene_list}{list, gene symbols as names and `stat` or `p-value` as values, decreasing
arranged. will overrite the diff as input}

\item{set}{set in c('all','all3','gsea','hm','kegg','im','im_jnj') 'all' - all the gmt
sets in the  refs, currently: Hallmark, KEGG, Immune, Immune2(im_jnj) 'all3' - hm, kegg,
immune do the 3 gmt sets 'gsea' - hm, kegg do the 2 gmt sets 'hm','kegg','im','im_jnj'
use the corresponding single gmt set}

\item{p.adjust.t}{padj threshold, default 0.05}

\item{p.t}{pvalue threshold, default 0.01}

\item{...}{}

\item{log2fc.abs.t}{log2FC threshold, default 1}
}
\value{
a list obj of class 'y.GO.res', used for input for yplot_GO_res
}
\description{
GSEA analysis using different expression results
}
