% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ydo_ssGSEA}
\alias{ydo_ssGSEA}
\title{do_ssGSEA analysis}
\usage{
ydo_ssGSEA(
  expr_matrix,
  todo = list("hm", "kegg", "im", "im_jnj"),
  ref.gmt = NULL,
  ref.gmt.name = NULL,
  mx.diff = FALSE,
  verbose = FALSE,
  method = "ssgsea",
  kcdf = "Gaussian",
  parallel.sz = 1,
  ...
)
}
\arguments{
\item{expr_matrix}{a numeric matrix witch rows are genes and columns are patients}

\item{todo}{list, combinations of elements in list('hm','kegg','im','im_jnj')
the collections todo ssGSEA, default run all 4 process, all the keys is the keys of parameter
<ref> in function `yload_list_gmt`}

\item{ref.gmt}{=NULL, if provided, must be a list like yload_list_gmt returns
or a file_path to a gmt.txt file like in yload_list_gmt <ref>, use it as <ref_gmt_list> todo ssGSEA}

\item{ref.gmt.name}{=NULL, if `ref.gmt` is provided, this will be the key-name of the returned results list
if is NULL, the returned list is unamed}

\item{mx.diff, verbose, method, kcdf, parallel.sz, ...}{pass to GSVA::gsva, details see GSVA::gsva docs
please keep `method` 'ssgsea' to do ssGSEA analysis}
}
\value{
list
}
\description{
depends on yload_list_gmt
 use the GSVA::gsea method which need 2 critical parameters
 a <expr_matrix> is the numeric matrix of expression, normed count of DSeq2 or tpm
 a <ref_gmt_list> which is the returns of yload_list_gmt, see details of yload_list_gmt documentation
}
