% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ydumpto}
\alias{ydumpto}
\title{Title}
\usage{
ydumpto(
  x,
  fname = NULL,
  export = NULL,
  outputdir = OUTPUTROOT,
  ext = NULL,
  prefix = NULL,
  flag = NULL,
  worker = NULL,
  verbose = T,
  suffix = NULL,
  mkdir = F,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
