% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{yfile_path}
\alias{yfile_path}
\title{Return nice format path values of file.path}
\usage{
yfile_path(...)
}
\arguments{
\item{...}{parameters of path to join}
}
\value{
character represent the full path
}
\description{
join given list of folders, likely the R version of os.path.join
may cause some unpredictable return values at certain circumstances
}
