% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{yfunc_args}
\alias{yfunc_args}
\title{Get the formal args of a function name, Only used inside a function}
\usage{
yfunc_args(func = NULL, get_default_values = FALSE)
}
\arguments{
\item{func}{c(character,NULL,symbol), character/symbol represents a function, if NULL get the formal args of the
parent function}

\item{get_default_values}{FALSE, only used when func is NULL, also get the default values of a given function?}
}
\value{
character vector if get_default_values==FALSE NOT recommended to use, cause can not determine the exact
  behavior at current at 2023-04-19 seems to get the list with names are the arg names and values are the default
  values at the run time
}
\description{
Get the formal args of a function name, Only used inside a function
}
\examples{
\dontrun{
  yfunc_args(utils::read.table)
}

}
