% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ygen_anno_color}
\alias{ygen_anno_color}
\title{ygen_anno_color}
\usage{
ygen_anno_color(
  df,
  palette = NULL,
  colors = NULL,
  cols = c("Clin_classification"),
  .autoNAcolor = TRUE,
  .assign_global = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{palette}{color palette, provide HEX color string vectors or use pal_*()(MAX_COLOR_NUM) to get a palette
if palette is NULL, use palette = pal_png()(10)}

\item{colors}{a list of index of colors in palette `colors` is a list with the same length of cols,
each element will act on the corresponding element of cols, if colors is NULL, random colors will be used, the funcion will invoke yset_random_anno_colors
for each element of colors, if  NA or NULL, use palette(1:length()) to generate colors,
 if 'random', use sample(palette, length()) to generate colors,
 if a integer vector, use palette$color to generate colors,}

\item{cols}{generate colors for these columns in  `df`}

\item{.autoNAcolor}{if TRUE, set color of NA value to '#BBBBBB'}

\item{.assign_global}{assign to global env or not, if colors is NULL, .assign_global will be set to FALSE}
}
\value{
list of colors
}
\description{
random colors for annotation columns (set `colors=NULL`)or set with a list of index of colors in palette
 used for generate COL_ANNO for maftools::oncoplot(annotationColor = COL_ANNO)
}
\examples{
df = data.frame(a = c('a','a','b'), b = c('a','c','c'), c = c('a','b','c'))
ygen_anno_color(df, cols = c('a','b','c'),)
}
