% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ygen_subMafs_}
\alias{ygen_subMafs_}
\title{SubsetMaf maftools::maf object}
\usage{
ygen_subMafs_(
  maf,
  glx,
  col = NULL,
  id_col = "Tumor_Sample_Barcode",
  .container = NULL,
  .assignGlobal = TRUE,
  ...
)
}
\arguments{
\item{maf}{maftools::maf object}

\item{glx}{group data used to subset the maf obj}

\item{col}{group column name used to subset the maf}

\item{id_col}{id_col used to link glx and maf Tumor_Sample_Barcode field}

\item{...}{}
}
\value{
if .assignGlobal==FALSE, return list else return NULL
}
\description{
SubsetMaf maftools::maf object
}
