% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ygenerate_annotation_colors}
\alias{ygenerate_annotation_colors}
\title{Generate color palette for annotation}
\usage{
ygenerate_annotation_colors(
  annotation,
  annotation_colors = NA,
  palette = "npg",
  show_all_ggsci_palette = FALSE
)
}
\arguments{
\item{show_all_ggsci_palette}{}
}
\description{
Generate color palette for annotation
}
