% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_signatures.R
\name{yget_known_ref_cancer_signatures}
\alias{yget_known_ref_cancer_signatures}
\title{Get pre-defined Cosmic Signatures by version}
\usage{
yget_known_ref_cancer_signatures(cosmic_ver = "v2")
}
\arguments{
\item{cosmic_ver}{{v2,v3}}
}
\value{
matrix, rows are 96 tri-nucleotides mutation types, columns are signatures
}
\description{
Get pre-defined Cosmic Signatures by version
}
