% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_signatures.R
\name{yget_maftools_signatures}
\alias{yget_maftools_signatures}
\title{Title}
\usage{
yget_maftools_signatures(
  laml_or_tnm,
  rank = NA,
  cosmic_ver = "v2",
  plot = FALSE,
  search_rank = FALSE,
  nTry = 6
)
}
\arguments{
\item{nTry}{}
}
\description{
Title
}
