% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_signatures.R
\name{yget_signatures_related_all}
\alias{yget_signatures_related_all}
\title{extract tnm, signatures and signatures contributions using maf objects}
\usage{
yget_signatures_related_all(
  laml,
  cosmic_ver = "v2",
  rank = 3,
  nrun = 3,
  single_core = TRUE,
  ...
)

yget_signatures_related_all(
  laml,
  cosmic_ver = "v2",
  rank = 3,
  nrun = 3,
  single_core = TRUE,
  ...
)
}
\arguments{
\item{laml}{maftools::maf object}

\item{cosmic_ver}{c('v2','v3')}

\item{rank}{number of signatures extracted}

\item{nrun}{resample times}

\item{single_core}{[MutationalPatterns::extract_signatures()]}

\item{...}{pass to MutationalPatterns::extract_signatures}
}
\value{
list, of list(mut_mat=mut_mat,extracted_n_signatures_suit=custom_sigs,contribution=fit_res$contribution,tnm=laml.tnm)

list contains:
 mut_mat: the tri-nucleotide-matrix:TNM
 extracted_n_signatures_suit: extracted signatures
 contribution: matrix of each samples' known COSMIC signatures' contribution components
 tnm: list of maftools::trinucleotideMatrix return values
}
\description{
using a maftools::readMaf returned MAF object
sequentially use maftools::trinucleotideMatrix ->
MutationalPatterns::extract_signatures ->
MutationalPatterns::fit_to_signatures to generate cosmic signatures
cosmic_ver, v2 means 30 v2 COSMIC signatures, v3 means the 79 v3 SBS signatures
rank, nrun and single_corepassed to MutationalPatterns::extract_signatures
}
\details{
use maftools::trinucleotideMatrix to get tri-nucleotides mutations
use MutationalPatterns::extract_signatures to extract maf specific signatures
use MutationalPatterns::fit_to_signatures to get cosine similaries and contributions
}
