% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ygetlast}
\alias{ygetlast}
\title{Get last element(s)}
\usage{
ygetlast(.obj, keys = NULL, .squeeze = FALSE, .retmode = "list")
}
\arguments{
\item{.obj}{the iterable, list}

\item{keys}{keys must be a vector of keys used for searching in names(.obj<list>)}

\item{.squeeze}{boolean,  .squeeze = T equals .retmode == 'squeeze'}

\item{.retmode}{in c('list','squeeze'), 'list' is default}
}
\value{
last object in list .obj
}
\description{
Get the last element which match the keys, of an iterable, like ypop, but do not change .obj,
*** Usually used for screening out unused arguments passed by ... in a function ***
Used by .ydumpto.function
}
\details{
Caution: list(1,3,4) |> names() -> NULL
list(1,3,a=4) |> names() -> c('','','a')
}
