% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{yhas_colnames}
\alias{yhas_colnames}
\title{Check DataFrame rownames and column names}
\usage{
yhas_colnames(df)
}
\arguments{
\item{df}{must be a class of data.frame}
}
\value{
TRUE or FASLE
}
\description{
like [yhas_rownames()] check whether a data.frame has rownames or not TRUE if there is a rownames vector else FALSE
}
