% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{yhas_rownames}
\alias{yhas_rownames}
\title{Check whether a DataFrame has true rownames}
\usage{
yhas_rownames(df)
}
\arguments{
\item{df}{must be an obj whose class is data.frame}
}
\value{
TRUE or FALSE
}
\description{
Check whether a DataFrame has true rownames
}
\details{
check whether a data.frame has rownames or not TRUE if there is a rownames vector not like 1:N (The natural one) else FALSE 一般用来确定diff expr等data.frame是否有symbol作为rownames，还是没有rownames，symbol为data.frame的一列
}
