% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yinfer_group_col}
\alias{yinfer_group_col}
\title{guess group column name in the data.frame}
\usage{
yinfer_group_col(df, raiseError = TRUE)
}
\arguments{
\item{df}{}
}
\description{
guess group column name in the data.frame
}
