% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{yis_valid_unreserved}
\alias{yis_valid_unreserved}
\title{check string is a valid R var/obj name}
\usage{
yis_valid_unreserved(string)
}
\arguments{
\item{string}{}
}
\value{
TRUE, if string is a valid else FALSE
}
\description{
check string is a valid R var/obj name, return TRUE if string is valid False if it is not
}
\examples{
yis_valid_unreserved('.jjj') # returns TRUE
yis_valid_unreserved('_jjj') # returns FALSE
}
