% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{yload_gmt}
\alias{yload_gmt}
\title{load symbol.gmt files using [clusterProfile::read.gmt()] into data.frame format obj}
\usage{
yload_gmt(key = NULL, base_ref_dir = NULL)
}
\arguments{
\item{key}{one of names(`ref`), default one of c('hm','kegg','im','im_jnj'),}

\item{base_ref_dir}{=NULL, if NULL use global SSGSEA_BASE_REF_DIR value
search gmt files in the base_ref_dir folder}

\item{ref}{key=filename pairs list, the `key` is search in. No need to change}
}
\value{
data.frame with first column is term, and second columns is gene, like
                              term    gene
1 HALLMARK_TNFA_SIGNALING_VIA_NFKB    JUNB
2 HALLMARK_TNFA_SIGNALING_VIA_NFKB   CXCL2
3 HALLMARK_TNFA_SIGNALING_VIA_NFKB    ATF3
4 HALLMARK_TNFA_SIGNALING_VIA_NFKB  NFKBIA
5 HALLMARK_TNFA_SIGNALING_VIA_NFKB TNFAIP3
...
}
\description{
see docs of yload_list_gmt, yload_gmt
}
\details{
The source of the gmt file is:
hm = 'h.all.v7.5.symbols.gmt',
kegg = 'c2.cp.kegg.v7.5.symbols.gmt',
im = 'immune.cell.rep.symbols.gmt',
im_jnj = 'immune.jnj.immunity.symbols.gmt'

ref = list( hm='hm.df', kegg='kegg.df', im='im.df',im_jnj='im_jnj.df')
}
\examples{
\dontrun{
gmt = yload_gmt('hm')
}

}
