% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{yload_gmtfile2list}
\alias{yload_gmtfile2list}
\title{read gmt file as a list}
\usage{
yload_gmtfile2list(gmtfile)
}
\arguments{
\item{gmtfile}{a gmt file. Examples are from MSigDB Collections.
A list of gene set could be find in the vignette of cogena}
}
\value{
a gmt list
}
\description{
read Gene Matrix Transposed (gmt) file and output a list with the the first
column as the names of items in the list. see
\href{http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats}{Gene Matrix Transposed file format}
for more details.
}
\examples{
\dontrun{
anno <- "c2.cp.kegg.v7.01.symbols.gmt.xz"
gmtfile <- system.file("extdata", anno, package="cogena")
gl <- gmt2list(gmtfile)
}

}
\seealso{
yload_gmt

gmtlist2file
}
