% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yload_laml_maf}
\alias{yload_laml_maf}
\title{Title}
\usage{
yload_laml_maf(maf_ = NULL, gl_ = NULL, laml = NULL, only_12_cols = TRUE, ...)
}
\arguments{
\item{maf_}{a MAF style table file name}

\item{gl_}{a group list file name}

\item{laml}{a maftools::MAF object}

\item{only_12_cols}{only use first 12 columns of read maf_ table (maftools standard input columns numbers)}

\item{...}{}
}
\value{
maftools::MAF object
}
\description{
Title
}
