% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{yload_list_gmt}
\alias{yload_list_gmt}
\title{load gmt files using read.table into list format obj}
\usage{
yload_list_gmt(key = NULL, base_ref_dir = NULL)
}
\arguments{
\item{key}{one of names(`ref`), default one of c('hm','kegg','im','im_jnj')}

\item{base_ref_dir}{=NULL, if NULL use global SSGSEA_BASE_REF_DIR value}

\item{ref}{key=filename pairs list, the `key` is search in. No need to change}

\item{gmt_txt_file_path}{=NULL, a file path string, file must the same as the files listed in the `ref`
if provided, read the file of the gmt_txt_file_path, ignore all other parameters}
}
\value{
list with names is term, and value is corresponding genes, like
 list(HALLMARK_ADIPOGENESIS=c('FABP4','ADIPOQ','PPARG',...),
     HALLMARK_ALLOGRAFT_REJECTION=c('PTPRC','IL12B','TGFB1','IL12A','CD3E',...),
     HALLMARK_ANDROGEN_RESPONSE=c('KLK3','KLK2','ACSL3','PIAS1','CAMKK2',...),
     HALLMARK_ANGIOGENESIS=c('VCAN','POSTN','FSTL1',...),
     ...)
}
\description{
see docs of yload_list_gmt, yload_gmt
}
\seealso{
yload_gmt
}
