% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yload_symbols_cancer_related_genes_1190}
\alias{yload_symbols_cancer_related_genes_1190}
\title{Load GC curated cancer-related gene symbols}
\usage{
yload_symbols_cancer_related_genes_1190()
}
\value{
a character vector
}
\description{
Load GC curated cancer-related gene symbols
}
\examples{
cr = yload_symbols_cancer_related_genes_1190()
}
