% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ylog}
\alias{ylog}
\title{Log messages}
\usage{
ylog(..., verbose = TRUE, .addtime = TRUE, .dev = 2)
}
\arguments{
\item{...}{parameters of string, which will be concat to print to}

\item{verbose}{if verbose is FALSE, stop logging anything}

\item{.addtime}{add timestamp at the front when print}

\item{.dev}{in [1,2] default 2(stderr), .dev==2 mean stderr, .dev==1 means stdout}
}
\value{
Nothing is returned
}
\description{
Log messages with timestamps to the stdout or stderr
}
\examples{
\dontrun{
file = 'abc.txt'
dir = '/path/to/save'
ylog(file,'is written at',dir)
}

}
