% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yloop}
\alias{yloop}
\title{apply func over a list x}
\usage{
yloop(x, func)
}
\arguments{
\item{x}{the list to be itered}

\item{func}{apply the func over x items, func could be 1,2,3 parameters
1: func(value)
2: func(value,key)
3: func(value,key,index)}
}
\value{
Nothing is returned
}
\description{
apply func over a list x
}
