% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{ynormalize_count_bydeseq2}
\alias{ynormalize_count_bydeseq2}
\title{Title use DESeq2 to do normalization and/or diff expr analysis for count data}
\usage{
ynormalize_count_bydeseq2(
  count_file_,
  group_file_,
  cnt_,
  glx_,
  frm = "export/",
  levels = NULL,
  export = F,
  outputdir = "export/",
  norm = T,
  diff = T
)
}
\arguments{
\item{count_file_}{you either provide @count_file_ which refer to a file path which will be read by `yload_dfx` or @cnt_ a data.frame/matrix obj containing rna count data with rows are genes and columns are patients}

\item{group_file_}{you either provide @group_file_ which refer to a file path which will be read by `yload_dfx` or @glx_ which is a data.frame/matrix obj contaning group design information}

\item{cnt_}{see @count_file_}

\item{glx_}{see @group_file_}

\item{frm}{will be pass to `yload_dfx` if you pass @count_file_ / @group_file_}

\item{levels}{which levels will be compared, if NULL or NA use the first 2 levels sorted alphabetically in Clin_classification column of @glx_}

\item{export}{output data results, default FALSE}

\item{outputdir}{set the output data directory}

\item{norm}{default TRUE, boolean, whether to do normalization}

\item{diff}{default TRUE, boolean, whether to do diff expression analysis}
}
\value{
A data.frame, the DESeq2 different expression analysis output data.frame
The DESeq2 different expression analysis output contains several columns.
The “baseMean” column contains the mean of normalized counts for each gene across all samples.
The “log2FoldChange” column contains the log2 fold change between two groups1.
The “lfcSE” column contains the standard error of the log2 fold change1.
The “stat” column contains the Wald test statistic. The Wald test is commonly used for hypothesis testing when comparing two groups.
   And “stat” column is determined by the log2FoldChange and pvalue column values which makes it a good ranking index for further order based analysis like GSEA pre-ranked procedure.
The “pvalue” column contains the p-value of the Wald test.
The “padj” column contains the adjusted p-value for multiple testing correction.
}
\description{
Title use DESeq2 to do normalization and/or diff expr analysis for count data
}
