% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{yplot_GSEA_res}
\alias{yplot_GSEA_res}
\title{Title}
\usage{
yplot_GSEA_res(
  res,
  .TOP = 30,
  figureTitle = "GSEA of DEGs",
  .lineheight = 0.75,
  .T_ylab_wrap = 40,
  .legends.color.title = "Expression",
  .legends.color.labels = c("Down-regulated", "Up-regulated"),
  ...
)
}
\arguments{
\item{.TOP}{plot top N features}

\item{.T_ylab_wrap}{the width of left y-axis labels}
}
\description{
Title
}
