% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_heatmap}
\alias{yplot_heatmap}
\title{Plot heatmap using ComplexHeatmap}
\usage{
yplot_heatmap(
  mat,
  colData = NULL,
  annotation_colors = NULL,
  scale = "auto",
  by_group = TRUE,
  params_top_annotation = NULL,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  clustering_method_columns = "ward.D2",
  clustering_method_rows = "ward.D2",
  name = "Z-Score",
  row_title = NA,
  show_row_dend = TRUE,
  show_column_dend = TRUE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  color = "-2~2",
  column_split = NULL,
  heatmap_legend_param = NULL,
  ...
)
}
\arguments{
\item{mat}{matrix, usually columns is tsb, if not matrix, will try to coerce using [as.matrix()]}

\item{colData}{None, or the data.frame with rownames (is tsb) must be accord with colnames(`mat`) and columns of
colData will be draw at the top annotation of Heatmap, each column a line}

\item{annotation_colors}{define the colors used for draw top annotation, representing colData}

\item{scale}{default 'auto', use [yscale_rows()] to standardize each rows}

\item{by_group}{if colData provided}

\item{params_top_annotation}{plotting params to draw top annotation}

\item{cluster_rows}{pass to [ComplexHeatmap::Heatmap()]}

\item{cluster_columns}{pass to [ComplexHeatmap::Heatmap()]}

\item{clustering_method_columns}{pass to [ComplexHeatmap::Heatmap()]}

\item{clustering_method_rows}{pass to [ComplexHeatmap::Heatmap()]}

\item{name}{pass to [ComplexHeatmap::Heatmap()]}

\item{row_title}{pass to [ComplexHeatmap::Heatmap()]}

\item{show_row_dend}{pass to [ComplexHeatmap::Heatmap()]}

\item{show_column_dend}{pass to [ComplexHeatmap::Heatmap()]}

\item{show_row_names}{pass to [ComplexHeatmap::Heatmap()]}

\item{show_column_names}{pass to [ComplexHeatmap::Heatmap()]}

\item{color}{in format 'Num1~Num2',default '-2~2', will be converted to [circlize::colorRamp()] function then pass to
[ComplexHeatmap::Heatmap()] as col. if the two Nums span 0, then 0 will be inserted into the middle
ELSE you can provide a custom [circlize::colorRamp()] object}

\item{column_split}{pass to [ComplexHeatmap::Heatmap()]}

\item{heatmap_legend_param}{pass to [ComplexHeatmap::Heatmap()]}

\item{...}{other legal params pass to [ComplexHeatmap::Heatmap()]}
}
\value{
list
}
\description{
Plot heatmap using ComplexHeatmap
}
