% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_pca}
\alias{yplot_pca}
\title{PCA analysis for mat}
\usage{
yplot_pca(
  normd,
  colData,
  color_col = NULL,
  add_label = FALSE,
  add_polygon = FALSE,
  tag_fontsize = 6,
  max.overlaps = Inf,
  alpha = 0.6,
  ...
)
}
\arguments{
\item{normd}{data.fram or matrix, only columns will be preserved, rows will be ruduced in dimensions}

\item{tag_fontsize}{pass to `size` in geom_label_repel}

\item{max.overlaps}{pass to ggrepel}

\item{alpha}{pass to geom_point}

\item{...}{}
}
\value{
list(gg = gg, pca = pca, data = data)
}
\description{
PCA analysis for mat
}
