% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_signatures.R
\name{yplot_signatrues_extracted_gg}
\alias{yplot_signatrues_extracted_gg}
\title{Title}
\usage{
yplot_signatrues_extracted_gg(
  laml.cos,
  laml.sig,
  titles = "",
  .with_ori_data = FALSE
)
}
\arguments{
\item{.with_ori_data}{}
}
\description{
Title
}
