% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{yplot_survival}
\alias{yplot_survival}
\title{Title}
\usage{
yplot_survival(
  data,
  covariates,
  key = NULL,
  test.method = "wald",
  tags = c("month", "status"),
  auto_anno = TRUE,
  join = "_"
)
}
\arguments{
\item{join}{}
}
\description{
Title
}
