% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_signatures.R
\name{yplot_tnm_cosmic}
\alias{yplot_tnm_cosmic}
\title{Title}
\usage{
yplot_tnm_cosmic(
  res.maftools.sigs,
  sel_sig = "1",
  sel_anno = "best",
  title = ""
)
}
\arguments{
\item{title}{}
}
\description{
Title
}
