% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_venns_bydf}
\alias{yplot_venns_bydf}
\title{Title}
\usage{
yplot_venns_bydf(
  df,
  id_col,
  group_col = "Clin_classification",
  .ret = FALSE,
  ...
)
}
\arguments{
\item{df}{}

\item{.ret}{if TRUE return the res object}
}
\value{
if .ret==TRUE return the res object else nothing is returned
}
\description{
Title
}
