% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_volcano_using_de}
\alias{yplot_volcano_using_de}
\title{Plot volcano plot}
\usage{
yplot_volcano_using_de(
  diff_expr,
  value_var = "padj",
  threshold = c(-2, 2),
  p = 0.05,
  add_label = FALSE,
  ...
)
}
\arguments{
\item{diff_expr}{RNA diff expr}

\item{value_var}{column name which used for filter}

\item{threshold}{vector of length 2, the log2FoldChange threshold}

\item{add_label}{whether to add the label text of genes}

\item{...}{}
}
\value{
list(gg=ggplot,de=data.frame)
}
\description{
Plot volcano plot
}
