% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypush}
\alias{ypush}
\title{ypush up-sert (upsert) items to list}
\usage{
ypush(iter, ..., .expand = TRUE, .as.whole = FALSE)
}
\arguments{
\item{iter}{an iterable obj like list}
}
\value{
updated `iter`
}
\description{
Unlike append, ypush is an update-append(like SQL upsert) version of append
}
\examples{
\dontrun{
list(a=1,b=2) \%>\% append(list(a=3,c=4)) # list(a=1,b=2,a=3,c=4)
list(a=1,b=2) \%>\% ypush(list(a=3,c=4)) # list(a=3,b=2,c=4)
list(a=1,b=2) \%>\% ypush(list(a=3,c=4),.expand=F) # list(a=3,b=2,c=4)
}
}
