% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypvalue_add_asterisks}
\alias{ypvalue_add_asterisks}
\title{Title}
\usage{
ypvalue_add_asterisks(pvalues)
}
\arguments{
\item{pvalues}{a numeric vector}
}
\value{
character, * formatted pvalues
}
\description{
Title
}
