% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yrmlast}
\alias{yrmlast}
\title{Remove the last matched by keys}
\usage{
yrmlast(.obj, keys = NULL, .byindex = FALSE)
}
\arguments{
\item{.obj}{is a named or unamed list or vector}

\item{keys}{is a unnamed vector}

\item{.byindex}{boolean}
}
\value{
the .obj with keys removed
}
\description{
yrmlast is the counterpart of the ygetlast, details See examples
}
\examples{
\dontrun{ x = list(a=4,b=3,c=2,d=1) # if .byindex == FALSE (default) # if .obj is named list or
vector, removal depents on keys, and `keys` is treated as names in .obj # if .obj is unnamed list
or vector, removal depents on values, and `keys` is treated as values in .obj x \%>\%
ygetlast(c('a','b','e')) # list(a=4,b=3) x \%>\% yrmlast(c('a','b','e')) # list(c=2,d=1) # if
`keys` is NULL,c(),list(), but not NA, remove nothing, `keys` must be an unnamed vector or list #
if .byindex == TRUE, keys must be positive numbers # representing the index of the .obj ranging
at 1~length(.obj) x \%>\% yrmlast(c(1,2),.byindex = T) # list(c=2,d=1) }
}
