% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ysave}
\alias{ysave}
\title{Save plots to pdf, and data.frame list to xlsx}
\usage{
ysave(
  fc = NULL,
  ...,
  table_title_lines = list(),
  outputdir = OUTPUTROOT,
  sep = "_",
  family = "sans"
)
}
\arguments{
\item{fc}{the obj to export}

\item{table_title_lines}{a list of title lines, element i will be the title line in the
corresponding table `fc[[i]]`, make sure the length of fc and table_title_lines are the same}

\item{outputdir}{output file into these dir}

\item{sep}{}

\item{family}{used in grDevices::cairo_pdf(family=family)}
}
\description{
auto save Robject into corresponding file format output file type is auto determined
  by the class(fc), currently support classes are 'ggplot', 'pheatmap','Heatmap' -> pdf files
  'call' -> eval(fc) -> pdf files 'data.frame', 'matrix' -> xlsx files 'list' of
  'data.frame'/'matrix's -> xlsx files with each element in the list -> a sheet in the xlsx
}
\examples{
\dontrun{
 data(mtcars)
 list(mtcars) \%>\% ysave('data','mtcars',outputdir='./') # file <'./data_mtcars.xlsx'> is created
 gg = ggplot(mtcars,aes(mpg,gear)) + geom_point()

 # file <'./mtcars.pdf'> which is 10 inches wide and 12 inches tall is created
 make.custom(10,12)
 gg \%>\% ysave('mtcars',outputdir='./')
}


}
