% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{yscale_rows}
\alias{yscale_rows}
\title{calculate Z-score of each row}
\usage{
yscale_rows(x)
}
\arguments{
\item{x}{matrix of number}
}
\value{
row Z-scored matrix
}
\description{
Z-score = (x-mean) / sd, calculate Z-score of each row
}
