% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{yset_random_anno_colors}
\alias{yset_random_anno_colors}
\title{Title}
\usage{
yset_random_anno_colors(
  df,
  palette = NULL,
  cols = c("Clin_classification"),
  .assign_global = FALSE
)
}
\arguments{
\item{.assign_global}{}
}
\description{
Title
}
