% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_utils.R
\name{yshow_available_pkgs_on_bioc}
\alias{yshow_available_pkgs_on_bioc}
\title{Show the available packages on Biocmanager}
\usage{
yshow_available_pkgs_on_bioc()
}
\value{
data.frame
}
\description{
Show the available packages on Biocmanager
}
