% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ysplit_file_name}
\alias{ysplit_file_name}
\title{Split a file name string into fname, extName}
\usage{
ysplit_file_name(str, fextname = NULL, mode = "auto", ...)
}
\arguments{
\item{fextname}{set extention name of}

\item{mode}{in ['auto','keep','add','replace'] default 'auto', behave, see example}

\item{...}{}
}
\value{
character of length 2
}
\description{
Split a file name string into fname, extName
}
