% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ysplit_path}
\alias{ysplit_path}
\title{split a string into a list of dirnames}
\usage{
ysplit_path(path)
}
\arguments{
\item{path}{}
}
\description{
if path start with /(root), the 1st element of returned list
is '/', multiple '/' like '//' will be treated as single '/'
}
