% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{ysubset_list_named}
\alias{ysubset_list_named}
\title{subset named list by keys}
\usage{
ysubset_list_named(x, keys)
}
\arguments{
\item{x}{is.list(x) must be TRUE}

\item{keys}{vector or list, return the element in x which match the keys, others in the keys are ignored.
if names(keys) is null, use values of keys
else use names(keys) to subset}
}
\value{
list with only names in intersect(names(x),~keys) is returned
}
\description{
subset the list x by names(keys) if keys is named list or vector
subset the list x by keys if keys is not named list or vector
}
\examples{
a = list(a=1,b=2,c=3)
k1 = c('a','c','d')
k2 = c(a="x",b=9,d='qq')
k3 = list(a='x',x='x2',c='ff')
ysubset_list_named(a,k1) # list(a=1,c=3)
ysubset_list_named(a,k2) # list(a=1,b=2)
ysubset_list_named(a,k3) # list(a=1,c=3)

}
