% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{yutils_wash_cnt_and_colData}
\alias{yutils_wash_cnt_and_colData}
\title{Title}
\usage{
yutils_wash_cnt_and_colData(
  cnt,
  colData,
  col.id = "Tumor_Sample_Barcode",
  col.group = "Clin_classification",
  levels = NULL,
  auto.filter = TRUE
)
}
\arguments{
\item{auto.filter}{}
}
\description{
Title
}
