% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ydo_GSEA}
\alias{ydo_GSEA}
\title{GSEA analysis using different expression results}
\usage{
ydo_GSEA(
  gene_list = NULL,
  set = "gsea",
  minGSSize = 10,
  maxGSSize = 500,
  p.adjust.t = 0.05,
  p.t = 0.2,
  pvalueCutoff = 1,
  ...
)
}
\arguments{
\item{gene_list}{list, gene symbols as names and `stat` or `p-value` as values, decreasing
arranged. will overrite the diff as input}

\item{set}{set in c('all','all3','gsea','hm','kegg','im','im_jnj') 'all' - all the gmt sets in
the  refs, currently: Hallmark, KEGG, Immune, Immune2(im_jnj) 'all3' - hm, kegg, immune do the
3 gmt sets 'gsea' - hm, kegg do the 2 gmt sets 'hm','kegg','im','im_jnj' use the corresponding
single gmt set}

\item{minGSSize}{will be passed to clusterProfiler::GSEA()}

\item{maxGSSize}{will be passed to clusterProfiler::GSEA()}

\item{p.adjust.t}{manual filter for GSEA result table, applied to FDR (`p.adjust`) column}

\item{p.t}{manual filter for GSEA result table, applied to p value (`pvalue`) column}

\item{pvalueCutoff}{passed into  clusterProfiler::GSEA(), controls the filter of the function,set
to 1 to avoid filter at this step.}

\item{...}{will be passed to  clusterProfiler::GSEA()}
}
\value{
a list obj of class 'y.GO.res', used for input for yplot_GO_res
}
\description{
GSEA analysis using different expression results
}
