% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ydo_GSEA}
\alias{ydo_GSEA}
\title{GSEA analysis using different expression results}
\usage{
ydo_GSEA(
  gene_list = NULL,
  set = "gsea",
  minGSSize = 10,
  maxGSSize = 500,
  p.adjust.t = 0.05,
  p.t = 0.2,
  .lineheight = 0.75,
  .legends.color.title = "Expression",
  .legends.color.labels = c("Down-regulated", "Up-regulated"),
  ...
)
}
\arguments{
\item{gene_list}{list, gene symbols as names and `stat` or `p-value` as values, decreasing
arranged. will overrite the diff as input}

\item{set}{set in c('all','all3','gsea','hm','kegg','im','im_jnj') 'all' - all the gmt
sets in the  refs, currently: Hallmark, KEGG, Immune, Immune2(im_jnj) 'all3' - hm, kegg,
immune do the 3 gmt sets 'gsea' - hm, kegg do the 2 gmt sets 'hm','kegg','im','im_jnj'
use the corresponding single gmt set}

\item{minGSSize}{will be passed to clusterProfiler::GSEA()}

\item{maxGSSize}{will be passed to clusterProfiler::GSEA()}

\item{p.adjust.t, p.t}{used to filter significant GSEA results from table `res.gsea@result`, p.adjust.t will be applied to p.adjust(FDR) values, and p.t will be pvalue}

\item{.lineheight}{yaxis label line heights}

\item{.legends.color.title}{the}

\item{.legends.color.labels}{the labels of the color legend, length 2 character vector, apply to values c(-1,1)}

\item{...}{will be passed to  clusterProfiler::GSEA()}
}
\value{
a list obj of class 'y.GO.res', used for input for yplot_GO_res
}
\description{
GSEA analysis using different expression results
}
