% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_GO_res}
\alias{yplot_GO_res}
\title{Plot the res obj of ydo_GO*}
\usage{
yplot_GO_res(
  go.res,
  go.tb.filter = "top10",
  .T_ylab_wrap = 50,
  ggsci_col_scheme = NULL,
  ...
)
}
\description{
Plot the res obj of ydo_GO*
}
