% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_GSEA_res}
\alias{yplot_GSEA_res}
\title{add plots to y.GSEA.resx objects}
\usage{
yplot_GSEA_res(
  res,
  p.adjust.t = NULL,
  p.t = NULL,
  .TOP = 30,
  figureTitle = "GSEA of DEGs",
  .lineheight = 0.75,
  .T_ylab_wrap = 40,
  .legends.color.title = "Expression",
  .legends.color.labels = c("Down-regulated", "Up-regulated"),
  ...
)
}
\arguments{
\item{res}{an 'y.GSEA.resx' object, generated by ydo_GSEA()}

\item{p.adjust.t, p.t}{if you want to re-filter the enrichement results with different thresholds
of FDR and pvalue thresholds, set the parameters to Non-NULL values}

\item{.TOP}{plot top N features, default 30}

\item{figureTitle}{}

\item{.lineheight}{yaxis label line heights}

\item{.T_ylab_wrap}{the width of left y-axis labels}

\item{.legends.color.title}{the title in legend "direction" which is the legend title of the
.legends.color.labels}

\item{.legends.color.labels}{the labels of the color legend, length 2 character vector, apply to
values c(-1,1)}
}
\description{
add plots to y.GSEA.resx objects
}
