% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_pca}
\alias{yplot_pca}
\title{PCA analysis for mat}
\usage{
yplot_pca(
  normd,
  colData,
  color_col = "Group",
  add_label = FALSE,
  tag_fontsize = 6,
  max.overlaps = Inf,
  alpha = 0.6,
  ...
)
}
\arguments{
\item{normd}{data.fram or matrix, only columns will be preserved, rows will be ruduced in dimensions}

\item{...}{}
}
\description{
PCA analysis for mat
}
