% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{yplot_tsne}
\alias{yplot_tsne}
\title{t-SNE analysis for mat}
\usage{
yplot_tsne(
  normd = NULL,
  colData = NULL,
  mat = NULL,
  color_col = NULL,
  add_label = FALSE,
  add_polygon = FALSE,
  perplexity = NA,
  dims = 2,
  ...
)
}
\arguments{
\item{normd}{data.fram or matrix, only columns will be preserved, rows will be ruduced in dimensions}

\item{...}{}
}
\description{
t-SNE analysis for mat
}
